pyproject.toml
[project]
name = "ibkr-web-api"
version = "0.1.0"
description = "Flask API for handling IBKR endpoints"
authors = [
    {name = "DanielGMorena",email = "danielgdelamorena@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.9,<3.14"
dependencies = [
    "requests (>=2.32.4,<3.0.0)",
    "pyyaml (>=6.0.2,<7.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "ib-insync (>=0.9.86,<0.10.0)",
    "fastapi (>=0.116.0,<0.117.0)",
    "pydantic-settings (>=2.10.1,<3.0.0)",
    "uvicorn (>=0.35.0,<0.36.0)",
    "types-pyyaml (>=6.0.12.20250516,<7.0.0.0)"
]



[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
package-mode = false

[tool.poetry.group.dev.dependencies]
pytest = "^8.4.1"
ruff = "^0.12.2"
pytest-cov = "^6.2.1"
pytest-asyncio = "^1.0.0"
httpx = "^0.28.1"
mypy = "^1.16.1"


[tool.poetry.group.dist.dependencies]
pyinstaller = "^6.14.2"

[tool.pytest.ini_options]
asyncio_mode = "auto"
addopts = "--cov=app --cov-report=term-missing"
testpaths = ["tests"]
pythonpath = ["."]
filterwarnings = [
    "ignore::DeprecationWarning:eventkit.*"
]

[tool.coverage.run]
omit = [
    "app/main.py",
    "app/run_server.py",
]

[tool.mypy]
python_version = "3.12"
plugins = ["pydantic.mypy"]
strict = true
ignore_missing_imports = true
warn_return_any = true
warn_unused_configs = true
files = ["app", "tests"]
exclude="tests"